requirements.txt
numpy==1.20.3
h5py==3.2.1
scipy==1.6.3
matplotlib==3.4.2
GitPython